add voting on proposals, update docs
