+++ bot/constants.py
@@ -1,11 +1,6 @@
 import os
 import logging
 
-"""
-######################################################################################################################################################
-Static & environment variables
-######################################################################################################################################################
-"""
 # Enable logging
 logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO)
 logger = logging.getLogger(__name__)
