+++ bot/helpers.py
@@ -1,10 +1,13 @@
 import json
 
 import requests
+from jigu.core.msg import MsgVote
 from telegram import InlineKeyboardButton, InlineKeyboardMarkup, TelegramError, KeyboardButton, ReplyKeyboardMarkup
 from requests.exceptions import RequestException
 from constants import *
-from service.governance_service import get_all_proposals_as_messages
+from messages import NETWORK_ERROR_MSG
+from service.governance_service import get_all_proposals_as_messages, get_active_proposals, get_proposal_by_id, \
+    jigu_proposal_to_text, get_my_vote, vote_on_proposal, is_wallet_provided
 
 """
 ######################################################################################################################################################
@@ -38,8 +41,8 @@ def show_my_nodes_menu_new_msg(context, chat_id):
 def show_governance_menu(context, chat_id):
     text = 'Click an option'
 
-    keyboard = [[InlineKeyboardButton("🗳 Show all proposals", callback_data='governance-all')],
-                [InlineKeyboardButton("Show active proposals and vote", callback_data='governance-active')]]
+    keyboard = [[InlineKeyboardButton("🗳 Show all proposals", callback_data='governance_all')],
+                [InlineKeyboardButton("🗳 ✅ Show active proposals and vote", callback_data='governance_active')]]
 
     try_message(context=context, chat_id=chat_id, text=text, reply_markup=InlineKeyboardMarkup(keyboard))
 
@@ -107,6 +110,105 @@ def on_show_all_proposals_clicked(update, context):
         try_message(context, query['message']['chat']['id'], message, reply_markup=None)
 
 
+def on_show_active_proposals_clicked(update, context):
+    query = update.callback_query
+    keyboard = [[]]
+
+    try:
+        active_proposals = get_active_proposals()
+    except Exception as e:
+        logger.error(e, exc_info=True)
+        try_message(context=context, chat_id=query['message']['chat']['id'], text=NETWORK_ERROR_MSG)
+        return
+
+    text = '🗳 ✅ **Active proposals**🗳 ✅ \n' \
+           'Click on any of the proposals to see details and vote options.\n\n'
+
+    if not active_proposals:
+        text += "No active proposals at the moment."
+
+    for proposal in active_proposals:
+        keyboard.append([InlineKeyboardButton(str(proposal.content.title), callback_data=f'proposal-{proposal.id}')])
+
+    try_message(context=context, chat_id=query['message']['chat']['id'], text=text,
+                reply_markup=InlineKeyboardMarkup(keyboard))
+
+
+def vote_on_proposal_details(update, _):
+    query = update.callback_query
+    proposal_id = query.data.split("-")[-1]
+    user_id = query.from_user['id']
+    try:
+        proposal = get_proposal_by_id(proposal_id)
+    except Exception as e:
+        logger.error(e)
+        query.edit_message_text(NETWORK_ERROR_MSG)
+        return
+
+    proposal_txt = jigu_proposal_to_text(proposal)
+    keyboard = [
+        [InlineKeyboardButton('⬅️ BACK', callback_data='governance_active')]]
+
+    message = ''
+
+    if not is_wallet_provided():
+        message = f"😢 *You can't vote, no MNEMONIC key provided.* 😢"
+    elif user_id not in ALLOWED_USER_IDS:
+        message = f'😢 *You are not allowed to vote because your id ({user_id}) is not whitelisted!* 😢'
+    else:
+        try:
+            my_vote = get_my_vote(proposal_id)
+        except Exception as e:
+            logger.error(e)
+            query.edit_message_text(NETWORK_ERROR_MSG)
+            return
+
+        if my_vote:
+            message = f'🎉 You voted *{my_vote}* 🎉'
+        else:
+            keyboard = [[
+                InlineKeyboardButton('✅ Yes', callback_data=f'vote-{proposal_id}-{MsgVote.YES}'),
+                InlineKeyboardButton('❌ No', callback_data=f'vote-{proposal_id}-{MsgVote.NO}'),
+            ], [
+                InlineKeyboardButton('❌❌ No with veto', callback_data=f'vote-{proposal_id}-{MsgVote.NO_WITH_VETO}'),
+                InlineKeyboardButton('🤷 Abstain', callback_data=f'vote-{proposal_id}-{MsgVote.ABSTAIN}'),
+            ], [InlineKeyboardButton('⬅️ BACK', callback_data='governance_active')]]
+
+    message += f"\n\n\n{proposal_txt}"
+
+    query.edit_message_text(message, parse_mode='markdown',
+                            reply_markup=InlineKeyboardMarkup(keyboard))
+
+
+def on_vote_clicked(update, _):
+    query = update.callback_query
+    _, proposal_id, vote = query.data.split("-")
+
+    keyboard = [[
+        InlineKeyboardButton('YES ✅', callback_data=f'vote_confirmed-{proposal_id}-{vote}'),
+        InlineKeyboardButton('NO ❌', callback_data=f'proposal-{proposal_id}')
+    ]]
+    text = f'⚠️ Do you really want to vote *{vote}* on proposal *{proposal_id}*? ⚠️'
+
+    query.edit_message_text(text, parse_mode='markdown', reply_markup=InlineKeyboardMarkup(keyboard))
+
+
+def vote_accept(update, _):
+    query = update.callback_query
+    _, proposal_id, vote = query.data.split("-")
+    keyboard = [[InlineKeyboardButton('⬅️ BACK', callback_data=f'proposal-{proposal_id}')]]
+
+    try:
+        vote_result = vote_on_proposal(proposal_id=proposal_id, vote_option=vote)
+        logger.info(f"Voted successfully. Transaction result:\n{vote_result}")
+    except Exception as e:
+        logger.error(e)
+        query.edit_message_text(NETWORK_ERROR_MSG, reply_markup=InlineKeyboardMarkup(keyboard))
+
+    query.edit_message_text(f"Successfully voted *{vote}* on proposal with id *{proposal_id}*", parse_mode='markdown',
+                            reply_markup=InlineKeyboardMarkup(keyboard))
+
+
 def show_confirmation_menu(update, text, keyboard):
     """
     "Are you sure?" - "YES" | "NO"
@@ -125,7 +227,8 @@ def send_message_to_all_platforms(context, chat_id, text):
 def send_slack_message(text):
     if SLACK_WEBHOOK:
         try:
-            requests.post(SLACK_WEBHOOK, data=json.dumps({'text': text}), headers={'Content-Type': 'application/json'})
+            requests.post(SLACK_WEBHOOK, data=json.dumps({'text': text}),
+                          headers={'Content-Type': 'application/json'})
         except RequestException as e:
             logger.error(f"Slack Webhook post request failed with:\n{e}")
