+++ bot/messages.py
@@ -1,1 +1,2 @@
-NETWORK_ERROR_MSG = '😱 There was an error while getting data 😱\nEither your internet connection or the node endpoint is down!'
+NETWORK_ERROR_MSG = '😱 There was an error while getting data 😱\nEither your internet connection or the node ' \
+                    'endpoint is down! '
